+ updated package structure + added more tests

+++ src/main/java/io/github/marbys/myrealworldapp/domain/service/CommentService.java
@@ -1,13 +1,13 @@
-package io.github.marbys.myrealworldapp.service;
+package io.github.marbys.myrealworldapp.domain.service;
 
-import io.github.marbys.myrealworldapp.repository.CommentRepository;
+import io.github.marbys.myrealworldapp.application.dto.CommentPostDTO;
 import io.github.marbys.myrealworldapp.domain.Article;
-import io.github.marbys.myrealworldapp.repository.ArticleRepository;
 import io.github.marbys.myrealworldapp.domain.Comment;
 import io.github.marbys.myrealworldapp.domain.User;
 import io.github.marbys.myrealworldapp.domain.model.CommentModel;
-import io.github.marbys.myrealworldapp.dto.CommentPostDTO;
-import io.github.marbys.myrealworldapp.repository.UserRepository;
+import io.github.marbys.myrealworldapp.infrastructure.repository.ArticleRepository;
+import io.github.marbys.myrealworldapp.infrastructure.repository.CommentRepository;
+import io.github.marbys.myrealworldapp.infrastructure.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -23,7 +23,7 @@ public class CommentService {
   private final UserRepository userRepository;
   private final ArticleRepository articleRepository;
 
-  public CommentModel addComment(String slug, CommentPostDTO commentPostDTO, long id) {
+  public Comment addComment(String slug, CommentPostDTO commentPostDTO, long id) {
     User author = userRepository.findById(id).orElseThrow(NoSuchElementException::new);
     Article article = articleRepository.findBySlug(slug).orElseThrow(NoSuchElementException::new);
 
@@ -31,15 +31,15 @@ public class CommentService {
     comment.setAuthor(author);
     comment.setArticle(article);
 
-    commentRepository.save(comment);
-    return CommentModel.fromComment(comment);
+    return commentRepository.save(comment);
   }
 
-  public void deleteComment(String slug, long commentId) {
+  public void deleteComment(String slug, long commentId, long id) {
+    User user = userRepository.findById(id).orElseThrow(NoSuchElementException::new);
     Comment comment =
         commentRepository.findById(commentId).orElseThrow(NoSuchElementException::new);
+    if (!comment.getAuthor().equals(user)) throw new IllegalStateException();
     Article article = articleRepository.findBySlug(slug).orElseThrow(NoSuchElementException::new);
-
     if (article.getComments().contains(comment)) commentRepository.deleteById(commentId);
   }
